Url.create can take string, URI, or Arroyo::Url now

+++ lib/arroyo/extensions.rb
@@ -2,7 +2,7 @@
 # @example
 #   4.MB
 #     => 4096
-class Integer
+class Numeric
   UNIT_BASE = 1024
   @@size_units = [
     { :name => 'KB', weight: UNIT_BASE },
